test: move runtime panic coverage out of source
